Fix linter and test executors and use temporary build directory

+++ packages/nx-sls/src/utils/file-utils.ts
@@ -58,5 +58,8 @@ export async function copyDirectory(src: PathLike, dest: PathLike) {
 }
 
 export async function deleteDirectory(path: PathLike) {
-    fs.rmdir(path, { recursive: true });
+    if (await exists(path)) {
+        return fs.rmdir(path, { recursive: true });
+    }
+    return;
 }
